backend/requirements.txt
alembic==1.14.0
asgiref==3.8.1
blinker==1.9.0
click==8.1.7
colorama==0.4.6
crispy-tailwind==1.0.3
distlib==0.3.9
Django==5.1.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
filelock==3.16.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.1
mysql-connector-python==9.1.0
mysqlclient==2.2.6
numpy==2.2.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
platformdirs==4.3.6
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.5.0
six==1.17.0
SQLAlchemy==2.0.36
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
